Split Context.h into one class per file
