+++ src/cudadev/CUDACore/ProduceContext.cc
@@ -1,0 +1,6 @@
+#include "CUDACore/ProduceContext.h"
+#include "CUDACore/cudaCheck.h"
+
+namespace cms::cuda {
+  void ProduceContext::commit() { cudaCheck(cudaEventRecord(event_.get(), stream())); }
+}
