+++ src/cudadev/CUDACore/ProductBase.h
@@ -10,7 +10,7 @@
 namespace cms {
   namespace cuda {
     namespace impl {
-      class Context;
+      class FwkContextBase;
 
       /**
        * The CUDA stream is shared between all the Event products of
@@ -66,13 +66,13 @@ namespace cms {
       int device() const { return device_; }
 
       // cudaStream_t is a pointer to a thread-safe object, for which a
-      // mutable access is needed even if the cms::cuda::Context itself
+      // mutable access is needed even if the ProductBase itself
       // would be const. Therefore it is ok to return a non-const
       // pointer from a const method here.
       cudaStream_t stream() const { return stream_->streamPtr().get(); }
 
       // cudaEvent_t is a pointer to a thread-safe object, for which a
-      // mutable access is needed even if the cms::cuda::Context itself
+      // mutable access is needed even if the ProductBase itself
       // would be const. Therefore it is ok to return a non-const
       // pointer from a const method here.
       cudaEvent_t event() const { return event_.get(); }
@@ -82,7 +82,7 @@ namespace cms {
           : stream_{std::move(stream)}, event_{std::move(event)}, device_{device} {}
 
     private:
-      friend class impl::Context;
+      friend class impl::FwkContextBase;
       friend class ProduceContext;
 
       // The following function is intended to be used only from Context
